requirements.txt
Flask==2.3.3
numpy==1.24.3
opencv-python==4.8.0.76
tensorflow==2.13.0
keras==2.13.1
pillow==10.0.0
werkzeug==2.3.7
requests==2.31.0
